+++ lib/src/responsive_layout.dart
@@ -1,13 +1,13 @@
 import 'package:flutter/widgets.dart';
 import 'package:responsive_mixin_layout/src/screen_size.dart';
 
-ScreenSize _screenSize(BuildContext context) {
-  final screenSize = context.findAncestorWidgetOfExactType<ScreenSize>();
-  if (screenSize == null) {
-    throw 'ResponsiveLayout must have an ancestor of type ScreenSize';
+ScreenSizes _screenSizes(BuildContext context) {
+  final screenSizes = context.findAncestorWidgetOfExactType<ScreenSizes>();
+  if (screenSizes == null) {
+    throw 'ResponsiveLayout must have an ancestor of type ScreenSizes';
   }
 
-  return screenSize;
+  return screenSizes;
 }
 
 /// Custom responsive layout widget
@@ -29,18 +29,18 @@ class ResponsiveLayout extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final screenSize = _screenSize(context);
+    final screenSizes = _screenSizes(context);
 
     return LayoutBuilder(
       builder: (context, constraints) {
-        if (mobile != null && constraints.maxWidth < screenSize.mobile) {
+        if (mobile != null && constraints.maxWidth < screenSizes.mobile) {
           return mobile!(context, constraints) ?? const Placeholder();
-        } else if (tablet != null && constraints.maxWidth < screenSize.tablet) {
+        } else if (tablet != null && constraints.maxWidth < screenSizes.tablet) {
           return tablet!(context, constraints) ?? const Placeholder();
         } else if (desktop != null &&
-            constraints.maxWidth < screenSize.desktop) {
+            constraints.maxWidth < screenSizes.desktop) {
           return desktop!(context, constraints) ?? const Placeholder();
-        } else if (tv != null && constraints.maxWidth < screenSize.tv) {
+        } else if (tv != null && constraints.maxWidth < screenSizes.tv) {
           return desktop!(context, constraints) ?? const Placeholder();
         }
 
@@ -61,18 +61,18 @@ mixin ResponsiveLayoutMixin<T extends Widget> on Widget {
   Widget? tvLayout(BuildContext context, BoxConstraints constraints) => null;
 
   Widget buildResponsiveLayout(BuildContext context) {
-    final screenSize = _screenSize(context);
+    final screenSizes = _screenSizes(context);
 
     return LayoutBuilder(
       builder: (context, constraints) {
         if (mobileLayout(context, constraints) != null &&
-            constraints.maxWidth < screenSize.mobile) {
+            constraints.maxWidth < screenSizes.mobile) {
           return mobileLayout(context, constraints) ?? const Placeholder();
         } else if (tabletLayout(context, constraints) != null &&
-            constraints.maxWidth < screenSize.tablet) {
+            constraints.maxWidth < screenSizes.tablet) {
           return tabletLayout(context, constraints) ?? const Placeholder();
         } else if (desktopLayout(context, constraints) != null &&
-            constraints.maxWidth < screenSize.desktop) {
+            constraints.maxWidth < screenSizes.desktop) {
           return desktopLayout(context, constraints) ?? const Placeholder();
         }
         return tvLayout(context, constraints) ?? const Placeholder();
@@ -94,18 +94,18 @@ mixin ResponsiveLayoutMixinStateful<T extends StatefulWidget> on State<T> {
   Widget? tvLayout(BuildContext context, BoxConstraints constraints) => null;
 
   Widget buildResponsiveLayout(BuildContext context) {
-    final screenSize = _screenSize(context);
+    final screenSizes = _screenSizes(context);
 
     return LayoutBuilder(
       builder: (context, constraints) {
         if (mobileLayout(context, constraints) != null &&
-            constraints.maxWidth < screenSize.mobile) {
+            constraints.maxWidth < screenSizes.mobile) {
           return mobileLayout(context, constraints) ?? const Placeholder();
         } else if (tabletLayout(context, constraints) != null &&
-            constraints.maxWidth < screenSize.tablet) {
+            constraints.maxWidth < screenSizes.tablet) {
           return tabletLayout(context, constraints) ?? const Placeholder();
         } else if (desktopLayout(context, constraints) != null &&
-            constraints.maxWidth < screenSize.desktop) {
+            constraints.maxWidth < screenSizes.desktop) {
           return desktopLayout(context, constraints) ?? const Placeholder();
         }
         return tvLayout(context, constraints) ?? const Placeholder();