+++ CHANGELOG.md
@@ -3,4 +3,9 @@
 * Initial release of the package. 
 * Added the  ResponsiveLayout  widget to create responsive layouts in Flutter. 
 * Provides custom widgets for mobile devices, tablets, desktops, and TVs. 
-* Allows customization of the application's appearance based on the device's screen size. 
+* Allows customization of the application's appearance based on the device's screen size.
+
+## 1.0.1
+
+* changed `ScreenSize` class name to `ScreenSizes`
+* updated extension on context to have getters